+++ packages/hash/frontend/src/pages/[shortname]/types/entity-type/[entity-type-id].page/definition-tab/property-list-card/shared/property-type-form-values.ts
@@ -14,8 +14,8 @@ interface PrimitiveExpectedValue {
 interface ArrayExpectedValue {
   typeId: "array";
   minItems: number;
-  maxItems?: number;
-  infinity?: boolean;
+  maxItems: number;
+  infinity: boolean;
   itemIds: string[];
 }
 
@@ -76,6 +76,7 @@ export const getDefaultExpectedValue = (
       typeId: "array",
       minItems: 0,
       maxItems: 0,
+      infinity: false,
       itemIds: [],
     };
   }
